Add PDF downloads to summary pages
Add the PDFMake and necessary objects to allow the summary page to be
downloaded as a a PDF from the Dox pages

Fix logic error

Non-index pages should have the 'PRINT' option.

Remove unused tests from WebPageGenerator script

Improvements to Print functionality

A good of set of testing files
------------------------------
1. Package: Lexicon Utility

2. SubFile: .111

3. Global: ^GMRD(120.57
4. Global: ^MDC(704.1161 (non-fileman)
5. Global: ^USR(8930.8

6. Routine: MCOBGC
7. Routine: XUOAAHL7
8. Routine: MCARAMLH

Add customization of the PDF download

Add the ability to select and deselect headers to include in the
"on-demand" PDF generation.  All data is selected by default.

Add PDF download for Package Interaction detail

Add a PDF download for the Package <-> Package interaction details page.

Expand the title list parser to look for an _1 or _2 to differentiate
between the sets of data in the page.

Change the Interaction details to capture all of its objects under a div
for each package.

Fix merge errors

Fix errors that were added in by the rebasing of the changes.

Change-Id: I1af2b34f8080a493c2298b4cc921b80aca13fe24